Extract Scrollable from useScroll
Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
